apps/meeting-assistant/pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "meeting-assistant"
version = "0.1.0"
description = "Real-time meeting assistant with RAG, LangChain and LangGraph"
requires-python = ">=3.10"

dependencies = [
    "fastapi>=0.104.0",
    "uvicorn>=0.24.0",
    "pydantic>=2.0.0",
    "pydantic-settings>=2.0.0",
    "python-dotenv>=1.0.0",
    "groq>=0.4.0",
    "langchain>=0.1.0",
    "langchain-groq>=0.0.1",
    "langchain-core>=0.1.0",
    "langgraph>=0.0.1",
    "langchain-pinecone>=0.0.1",
    "pinecone-client>=3.0.0",
    "websockets>=11.0.0",
    "aiohttp>=3.9.0",
    "python-multipart>=0.0.6",
    "sqlalchemy>=2.0.0",
    "psycopg2-binary>=2.9.0",
    "alembic>=1.13.0",
    "httpx>=0.25.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.0",
    "pytest-asyncio>=0.21.0",
    "black>=23.0.0",
    "ruff>=0.1.0",
    "mypy>=1.7.0",
]

[tool.black]
line-length = 100
target-version = ["py310"]

[tool.ruff]
line-length = 100
target-version = "py310"

[tool.mypy]
python_version = "3.10"
strict = true
